refactor: move image files to the relevant image folder
This moves the two image files that was in the components directory to the image directory.
