+++ public/static/images/contribution-flow/fees-on-top-illustration.png
