+++ public/static/images/contribution-flow/success-illustration.jpg
